+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Add "Search value on Flows" to taps
     - Add "Timeline tool" implementation
     - Add power stepping to stepping controls
+    - Add "Printer tool" implementation
     
 ### Changes